+++ pygeos/geometry.py
@@ -28,6 +28,8 @@ __all__ = [
     "get_rings",
     "get_precision",
     "set_precision",
+    "force_2d",
+    "force_3d",
 ]
 
 
@@ -118,7 +120,8 @@ def get_dimensions(geometry, **kwargs):
 def get_coordinate_dimension(geometry, **kwargs):
     """Returns the dimensionality of the coordinates in a geometry (2 or 3).
 
-    Returns -1 for not-a-geometry values.
+    Returns -1 for missing geometries (``None`` values). Note that if the first Z
+    coordinate equals ``nan``, this function will return ``2``.
 
     Parameters
     ----------
@@ -135,6 +138,8 @@ def get_coordinate_dimension(geometry, **kwargs):
     3
     >>> get_coordinate_dimension(None)
     -1
+    >>> get_coordinate_dimension(Geometry("POINT Z (0 0 nan)"))
+    2
     """
     return lib.get_coordinate_dimension(geometry, **kwargs)
 
@@ -735,3 +740,61 @@ def set_precision(geometry, grid_size, preserve_topology=False, **kwargs):
     """
 
     return lib.set_precision(geometry, grid_size, preserve_topology, **kwargs)
+
+
+@multithreading_enabled
+def force_2d(geometry, **kwargs):
+    """Forces the dimensionality of a geometry to 2D.
+
+    Parameters
+    ----------
+    geometry : Geometry or array_like
+    **kwargs
+        For other keyword-only arguments, see the
+        `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
+
+    Examples
+    --------
+    >>> force_2d(Geometry("POINT Z (0 0 0)"))
+    <pygeos.Geometry POINT (0 0)>
+    >>> force_2d(Geometry("POINT (0 0)"))
+    <pygeos.Geometry POINT (0 0)>
+    >>> force_2d(Geometry("LINESTRING (0 0 0, 0 1 1, 1 1 2)"))
+    <pygeos.Geometry LINESTRING (0 0, 0 1, 1 1)>
+    >>> force_2d(Geometry("POLYGON Z EMPTY"))
+    <pygeos.Geometry POLYGON EMPTY>
+    >>> force_2d(None) is None
+    True
+    """
+    return lib.force_2d(geometry, **kwargs)
+
+
+@multithreading_enabled
+def force_3d(geometry, z=0.0, **kwargs):
+    """Forces the dimensionality of a geometry to 3D.
+
+    2D geometries will get the provided Z coordinate; Z coordinates of 3D geometries
+    are unchanged (unless they are nan).
+
+    Parameters
+    ----------
+    geometry : Geometry or array_like
+    z : float or array_like, default 0.0
+    **kwargs
+        For other keyword-only arguments, see the
+        `NumPy ufunc docs <https://numpy.org/doc/stable/reference/ufuncs.html#ufuncs-kwargs>`_.
+
+    Examples
+    --------
+    >>> force_3d(Geometry("POINT (0 0)"), z=3)
+    <pygeos.Geometry POINT Z (0 0 3)>
+    >>> force_3d(Geometry("POINT Z (0 0 0)"), z=3)
+    <pygeos.Geometry POINT Z (0 0 0)>
+    >>> force_3d(Geometry("LINESTRING (0 0, 0 1, 1 1)"))
+    <pygeos.Geometry LINESTRING Z (0 0 0, 0 1 0, 1 1 0)>
+    >>> force_3d(None) is None
+    True
+    """
+    if np.isnan(z).any():
+        raise ValueError("It is not allowed to set the Z coordinate to NaN.")
+    return lib.force_3d(geometry, z, **kwargs)