+++ pygeos/tests/common.py
@@ -26,9 +26,9 @@ multi_polygon = pygeos.multipolygons(
 geometry_collection = pygeos.geometrycollections(
     [pygeos.points(51, -1), pygeos.linestrings([(52, -1), (49, 2)])]
 )
-point_z = pygeos.points(1.0, 1.0, 1.0)
-line_string_z = pygeos.linestrings([(0, 0, 0), (1, 0, 1), (1, 1, 2)])
-polygon_z = pygeos.polygons([(0, 0, 0), (2, 0, 1), (2, 2, 2), (0, 2, 3), (0, 0, 0)])
+point_z = pygeos.points(2, 3, 4)
+line_string_z = pygeos.linestrings([(0, 0, 4), (1, 0, 4), (1, 1, 4)])
+polygon_z = pygeos.polygons([(0, 0, 4), (2, 0, 4), (2, 2, 4), (0, 2, 4), (0, 0, 4)])
 geometry_collection_z = pygeos.geometrycollections([point_z, line_string_z])
 polygon_with_hole = pygeos.Geometry(
     "POLYGON((0 0, 0 10, 10 10, 10 0, 0 0), (2 2, 2 4, 4 4, 4 2, 2 2))"
@@ -36,9 +36,21 @@ polygon_with_hole = pygeos.Geometry(
 empty_point = pygeos.Geometry("POINT EMPTY")
 empty_point_z = pygeos.Geometry("POINT Z EMPTY")
 empty_line_string = pygeos.Geometry("LINESTRING EMPTY")
+empty_line_string_z = pygeos.Geometry("LINESTRING Z EMPTY")
 empty_polygon = pygeos.Geometry("POLYGON EMPTY")
 empty = pygeos.Geometry("GEOMETRYCOLLECTION EMPTY")
 line_string_nan = pygeos.linestrings([(np.nan, np.nan), (np.nan, np.nan)])
+multi_point_z = pygeos.multipoints([(0, 0, 4), (1, 2, 4)])
+multi_line_string_z = pygeos.multilinestrings([[(0, 0, 4), (1, 2, 4)]])
+multi_polygon_z = pygeos.multipolygons(
+    [
+        [(0, 0, 4), (1, 0, 4), (1, 1, 4), (0, 1, 4), (0, 0, 4)],
+        [(2.1, 2.1, 4), (2.2, 2.1, 4), (2.2, 2.2, 4), (2.1, 2.2, 4), (2.1, 2.1, 4)],
+    ]
+)
+polygon_with_hole_z = pygeos.Geometry(
+    "POLYGON Z((0 0 4, 0 10 4, 10 10 4, 10 0 4, 0 0 4), (2 2 4, 2 4 4, 4 4 4, 4 2 4, 2 2 4))"
+)
 
 all_types = (
     point,
@@ -78,6 +90,6 @@ def assert_geometries_equal(actual, expected):
     expected = np.broadcast_to(expected, actual.shape)
     mask = pygeos.is_geometry(expected)
     if np.any(mask):
-        assert pygeos.equals(actual[mask], expected[mask]).all()
+        assert pygeos.equals_exact(actual[mask], expected[mask]).all()
     if np.any(~mask):
         assert_array_equal(actual[~mask], expected[~mask])