+++ pygeos/tests/test_coordinates.py
@@ -7,7 +7,9 @@ from pygeos import apply, count_coordinates, get_coordinates, set_coordinates
 
 from .common import (
     empty,
+    empty_line_string_z,
     empty_point,
+    empty_point_z,
     geometry_collection,
     geometry_collection_z,
     line_string,
@@ -119,10 +121,10 @@ def test_get_coords_index_multidim(order):
 @pytest.mark.parametrize(
     "geoms,x,y,z",
     [
-        ([point, point_z], [2, 1], [3, 1], [np.nan, 1]),
-        ([line_string_z], [0, 1, 1], [0, 0, 1], [0, 1, 2]),
-        ([polygon_z], [0, 2, 2, 0, 0], [0, 0, 2, 2, 0], [0, 1, 2, 3, 0]),
-        ([geometry_collection_z], [1, 0, 1, 1], [1, 0, 0, 1], [1, 0, 1, 2]),
+        ([point, point_z], [2, 2], [3, 3], [np.nan, 4]),
+        ([line_string_z], [0, 1, 1], [0, 0, 1], [4, 4, 4]),
+        ([polygon_z], [0, 2, 2, 0, 0], [0, 0, 2, 2, 0], [4, 4, 4, 4, 4]),
+        ([geometry_collection_z], [2, 0, 1, 1], [3, 0, 0, 1], [4, 4, 4, 4]),
     ],
 )  # fmt: on
 def test_get_coords_3d(geoms, x, y, z, include_z):
@@ -239,6 +241,26 @@ def test_apply_check_shape():
 def test_apply_correct_coordinate_dimension():
     # ensure that new geometry is 2D with include_z=False
     geom = line_string_z
+    assert pygeos.get_coordinate_dimension(geom) == 3
+    new_geom = apply(geom, lambda x: x + 1, include_z=False)
+    assert pygeos.get_coordinate_dimension(new_geom) == 2
+
+
+@pytest.mark.parametrize("geom", [
+    pytest.param(empty_point_z, marks=pytest.mark.skipif(pygeos.geos_version < (3, 9, 0), reason="Empty points don't have a dimensionality before GEOS 3.9")),
+    empty_line_string_z,
+])
+def test_apply_empty_preserve_z(geom):
+    assert pygeos.get_coordinate_dimension(geom) == 3
+    new_geom = apply(geom, lambda x: x + 1, include_z=True)
+    assert pygeos.get_coordinate_dimension(new_geom) == 3
+
+
+@pytest.mark.parametrize("geom", [
+    pytest.param(empty_point_z, marks=pytest.mark.skipif(pygeos.geos_version < (3, 9, 0), reason="Empty points don't have a dimensionality before GEOS 3.9")),
+    empty_line_string_z,
+])
+def test_apply_remove_z(geom):
     assert pygeos.get_coordinate_dimension(geom) == 3
     new_geom = apply(geom, lambda x: x + 1, include_z=False)
     assert pygeos.get_coordinate_dimension(new_geom) == 2