+++ pygeos/tests/test_geometry.py
@@ -3,23 +3,32 @@ import pytest
 
 import pygeos
 
-from .common import all_types
+from .common import all_types, assert_geometries_equal
 from .common import empty as empty_geometry_collection
 from .common import (
     empty_line_string,
+    empty_line_string_z,
     empty_point,
+    empty_point_z,
     empty_polygon,
     geometry_collection,
+    geometry_collection_z,
     line_string,
     line_string_nan,
+    line_string_z,
     linear_ring,
     multi_line_string,
+    multi_line_string_z,
     multi_point,
+    multi_point_z,
     multi_polygon,
+    multi_polygon_z,
     point,
     point_z,
     polygon,
     polygon_with_hole,
+    polygon_with_hole_z,
+    polygon_z,
 )
 
 
@@ -175,16 +184,16 @@ def test_get_xyz_no_point(func, geom):
 
 
 def test_get_x():
-    assert pygeos.get_x([point, point_z]).tolist() == [2.0, 1.0]
+    assert pygeos.get_x([point, point_z]).tolist() == [2.0, 2.0]
 
 
 def test_get_y():
-    assert pygeos.get_y([point, point_z]).tolist() == [3.0, 1.0]
+    assert pygeos.get_y([point, point_z]).tolist() == [3.0, 3.0]
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 7, 0), reason="GEOS < 3.7")
 def test_get_z():
-    assert pygeos.get_z([point_z]).tolist() == [1.0]
+    assert pygeos.get_z([point_z]).tolist() == [4.0]
 
 
 @pytest.mark.skipif(pygeos.geos_version < (3, 7, 0), reason="GEOS < 3.7")
@@ -560,3 +569,73 @@ def test_empty():
     """Compatibility with empty_like, see GH373"""
     g = np.empty_like(np.array([None, None]))
     assert pygeos.is_missing(g).all()
+
+
+# corresponding to geometry_collection_z:
+geometry_collection_2 = pygeos.geometrycollections([point, line_string])
+empty_point_mark = pytest.mark.skipif(
+    pygeos.geos_version < (3, 9, 0),
+    reason="Empty points don't have a dimensionality before GEOS 3.9",
+)
+
+
+@pytest.mark.parametrize(
+    "geom,expected",
+    [
+        (point, point),
+        (point_z, point),
+        pytest.param(empty_point, empty_point, marks=empty_point_mark),
+        pytest.param(empty_point_z, empty_point, marks=empty_point_mark),
+        (line_string, line_string),
+        (line_string_z, line_string),
+        (empty_line_string, empty_line_string),
+        (empty_line_string_z, empty_line_string),
+        (polygon, polygon),
+        (polygon_z, polygon),
+        (polygon_with_hole, polygon_with_hole),
+        (polygon_with_hole_z, polygon_with_hole),
+        (multi_point, multi_point),
+        (multi_point_z, multi_point),
+        (multi_line_string, multi_line_string),
+        (multi_line_string_z, multi_line_string),
+        (multi_polygon, multi_polygon),
+        (multi_polygon_z, multi_polygon),
+        (geometry_collection_2, geometry_collection_2),
+        (geometry_collection_z, geometry_collection_2),
+    ],
+)
+def test_force_2d(geom, expected):
+    actual = pygeos.force_2d(geom)
+    assert pygeos.get_coordinate_dimension(actual) == 2
+    assert_geometries_equal(actual, expected)
+
+
+@pytest.mark.parametrize(
+    "geom,expected",
+    [
+        (point, point_z),
+        (point_z, point_z),
+        pytest.param(empty_point, empty_point_z, marks=empty_point_mark),
+        pytest.param(empty_point_z, empty_point_z, marks=empty_point_mark),
+        (line_string, line_string_z),
+        (line_string_z, line_string_z),
+        (empty_line_string, empty_line_string_z),
+        (empty_line_string_z, empty_line_string_z),
+        (polygon, polygon_z),
+        (polygon_z, polygon_z),
+        (polygon_with_hole, polygon_with_hole_z),
+        (polygon_with_hole_z, polygon_with_hole_z),
+        (multi_point, multi_point_z),
+        (multi_point_z, multi_point_z),
+        (multi_line_string, multi_line_string_z),
+        (multi_line_string_z, multi_line_string_z),
+        (multi_polygon, multi_polygon_z),
+        (multi_polygon_z, multi_polygon_z),
+        (geometry_collection_2, geometry_collection_z),
+        (geometry_collection_z, geometry_collection_z),
+    ],
+)
+def test_force_3d(geom, expected):
+    actual = pygeos.force_3d(geom, z=4)
+    assert pygeos.get_coordinate_dimension(actual) == 3
+    assert_geometries_equal(actual, expected)